# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  Backend_internal_models.User:
    properties:
      additional_notes:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      email_verification_token:
        type: string
      email_verified:
        type: boolean
      file_path:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      institution_name:
        type: string
      last_name:
        type: string
      major:
        type: string
      middle_name:
        type: string
      password:
        type: string
      password_reset_expires:
        type: string
      password_reset_token:
        type: string
      profile_picture:
        type: string
      role_id:
        type: integer
      student_id:
        type: string
      student_id_verification:
        type: string
      student_id_verified:
        type: boolean
      twofa_enabled:
        type: boolean
      twofa_image:
        type: string
      twofa_secret:
        type: string
      updated_at:
        type: string
      username:
        type: string
      year:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@pufacomputing.ac.id
    name: API Support
    url: https://compsci.president.ac.id/support
  description: Backend API for PUFA Computer Science with Candidates and Votes Management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://compsci.president.ac.id/terms
  title: PUFA Backend API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 'Login with username/email and password. Returns JWT token for
        authentication. Admin (role_id: 1) can create candidates. Computizen (role_id:
        2) with year 2025 can vote.'
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          properties:
            passcode:
              type: string
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with JWT token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid credentials or email not verified
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout the current user and invalidate the JWT token
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: 'Register a new user account. New users get Computizen role (role_id:
        2) by default. Only users with year 2025 can vote.'
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/Backend_internal_models.User'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request or validation error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"